requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
frozenlist==1.3.1
future==0.18.2
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.2
pycodestyle==2.9.1
requests==2.28.1
toml==0.10.2
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0
